move files, fix test
