+++ src/lib/parse.ts
@@ -1,4 +1,4 @@
-import { Trace } from './trace';
+import { Trace } from './misc/trace';
 import {
 	AssignMode,
 	BinaryOperator,
