+++ src/lib/run.ts
@@ -11,7 +11,7 @@ import {
 	SourceFile,
 	StatementNode,
 } from './ast';
-import { Trace } from './trace';
+import { Trace } from './misc/trace';
 
 const trace = Trace.getDefault().createChild(false);
